Exception handling - delete

+++ src/main/java/com/educandoweb/course/services/UserService.java
@@ -4,10 +4,12 @@ import java.util.List;
 import java.util.Optional;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.stereotype.Service;
 
 import com.educandoweb.course.entities.User;
 import com.educandoweb.course.repositories.UserRepository;
+import com.educandoweb.course.services.exceptions.DatabaseException;
 import com.educandoweb.course.services.exceptions.ResourceNotFoundException;
 
 @Service
@@ -15,28 +17,37 @@ public class UserService {
 
 	@Autowired
 	private UserRepository repository;
-	
-	public List<User> findAll(){
-		
+
+	public List<User> findAll() {
+
 		return repository.findAll();
 	}
-	
+
 	public User findById(Long id) {
 		Optional<User> user = repository.findById(id);
 		return user.orElseThrow(() -> new ResourceNotFoundException(id));
 	}
-	
-	public User insert (User user) {
+
+	public User insert(User user) {
 		return repository.save(user);
 	}
+
+	public void delete(Long id) {
+	       if (!repository.existsById(id)) {
+	            throw new ResourceNotFoundException(id);
+	        }
+	       
+	        try {
+	            repository.deleteById(id);
+	        } catch (DataIntegrityViolationException e) {
+	            throw new DatabaseException(e.getMessage());
+	        }
+	        
+	    }
 	
-	public void delete (Long id) {
-		repository.deleteById(id);
-	}
-	
-	public User update (Long id, User user) {
+	public User update(Long id, User user) {
 		User entity = repository.getReferenceById(id);
-		updateData (entity, user);
+		updateData(entity, user);
 		return repository.save(entity);
 	}
